% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all.R
\name{read_all}
\alias{read_all}
\title{Load a full MS Excel File}
\usage{
read_all(filename, environment = parent.frame(), verbose = TRUE)
}
\arguments{
\item{filename}{The path and the name to the MS Excel File.}

\item{environment}{Environment were the new variables will be located. By
default is parent environment.}

\item{verbose}{If set to \code{TRUE} usefull messages are shown.}
}
\description{
Given a MS Excel File it loads the full workbook creating as many variables as
sheets exists in the workbook.
}

