% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tab.R
\name{read_tab}
\alias{read_tab}
\title{Load a single tab of MS Excel File}
\usage{
read_tab(filename, tabname, environment = parent.frame(), verbose = TRUE)
}
\arguments{
\item{filename}{The path and the name to the MS Excel File.}

\item{tabname}{Name of the tab to be loaded.}

\item{environment}{Environment were the new variables will be located. By
default is parent environment.}

\item{verbose}{If set to \code{TRUE} usefull messages are shown.}
}
\description{
Given a MS Excel File and the name of a tab it loads the tab contentent
as variables.
}

