% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_all.R
\name{write_all}
\alias{write_all}
\title{Write objects to a MS Excel File}
\usage{
write_all(..., filename, verbose = TRUE)
}
\arguments{
\item{...}{\code{data.frame}s to be saved as sheets in a MS Excel File.}

\item{filename}{The path and the name to the MS Excel File.}

\item{verbose}{If set to \code{TRUE} usefull messages are shown.}
}
\description{
Given a collection of R objects (aka \code{data.frame}s), it saves them as
different sheets in a MS Excel File.
}

