% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sheet.R
\name{write_sheet}
\alias{write_sheet}
\title{Write single object to a MS Excel File}
\usage{
write_sheet(data, sheetname, filename, replace = FALSE, verbose = TRUE)
}
\arguments{
\item{data}{\code{data.frame} to be saved as sheet in a MS Excel File.}

\item{sheetname}{Name to be given to the sheet.}

\item{filename}{The path and the name to the MS Excel File.}

\item{replace}{If set to \code{TRUE} it replaces the content of the
given sheet with the current value of \code{data}.}

\item{verbose}{If set to \code{TRUE} usefull messages are shown.}
}
\description{
Given a single \code{data.frame}, it saves it as sheet in a MS Excel File.
}

